requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-cpf==0.1.0
django-environ==0.9.0
django-registration==3.3
django-templated-mail==1.1.1
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-registration==0.1.3
drf-spectacular==0.24.2
executing==1.1.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
model-bakery==1.9.0
mypy-extensions==0.4.3
oauthlib==3.2.2
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyrsistent==0.18.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
